NE-16: adiciona funcionalidade de desembolsos de empresas no cronograma

+++ PeD.Core/ApiModels/Cronograma/EmpresaCronogramaDto.cs
@@ -5,7 +5,7 @@ namespace PeD.Core.ApiModels.Cronograma
     public class EmpresaCronogramaDto
     {
         public string Nome { get; set; }
-        public List<int> Desembolso { get; set; }
-        public int Total { get; set; }
+        public List<double> Desembolso { get; set; }
+        public double Total { get; set; }
     }
 }
